crate-ai-rag/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
langchain>=0.0.350
chromadb>=0.4.18
sentence-transformers>=2.2.2
llama-cpp-python>=0.2.20
huggingface-hub>=0.19.4
torch>=2.0.0
pyyaml>=6.0.1
python-multipart>=0.0.6
watchdog>=3.0.0